+++ pymc/tests/test_smc.py
@@ -293,8 +293,8 @@ class TestSimulator(SeededTest):
 
         with self.SMABC_test:
             trace = pm.sample_smc(draws=1000, return_inferencedata=False)
-            pr_p = pm.sample_prior_predictive(1000)
-            po_p = pm.sample_posterior_predictive(trace, 1000)
+            pr_p = pm.sample_prior_predictive(1000, return_inferencedata=False)
+            po_p = pm.sample_posterior_predictive(trace, 1000, return_inferencedata=False)
 
         assert abs(self.data.mean() - trace["a"].mean()) < 0.05
         assert abs(self.data.std() - trace["b"].mean()) < 0.05